chore: add error info (#44)
This commit standardizes error handling in the `pttl` and `set`
functions, which were previously returning false on error rather than
using the `mapExceptionElseReturnFalse` function like all the other
functions. We also add the traceback from the Momento SDK error in the
`RedisException` error message so we can determine which API call
produced the error.

The basic example is also updated to configure the Momento client to use
5 gRPC channels instead of the default of 1.